chores: updated tests for zotero path
